requirements.txt
pandas
numpy==1.26.4
matplotlib
seaborn
scikit-learn
scipy
statsmodels
scikit-learn
pyarrow
holidays
fastapi
uvicorn
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
tensorflow
SQLAlchemy
pydantic-settings
